add analogous python example for benchmarking
