+++ rps_jax/example.py
@@ -1,57 +1,35 @@
-from rps_jax.robotarium import Robotarium
+import numpy as np
+from rps.robotarium import Robotarium
+from tqdm import tqdm
 
-import jax
-import jax.numpy as jnp
-import matplotlib.pyplot as plt
-from jax import jit, vmap, lax
+# Example usage
+def drive_in_circle(num_env, num_t):
+    for env in range(num_env):
+        # Create Robotarium object
+        r = Robotarium(number_of_robots=1, sim_in_real_time=False, show_figure=False, initial_conditions=np.array([[0.0, 0.0, 0]]).T)
 
+        # Define the angular velocity
+        radius = 1
+        v = 1
+        omega = v / radius
 
-class WrappedRobotarium(object):
-    def __init__(self, env, num_envs):
-        self.env = env
-        self.num_envs = num_envs
-    
-    def wrapped_step(self):
-        return self.env.step()
-    
-    def move_circle(self, pose, radius):
-        x, y, theta = pose
-        v = 1.0  # linear velocity
-        omega = v / radius  # angular velocity
-        return jnp.array([v, omega])[:, None]
+        poses = r.get_poses()
+        for t in tqdm(range(num_t)):
+            # Calculate the unicycle velocity commands
+            w = omega
 
-    def batched_step(self, poses, unused):
-        actions = vmap(self.move_circle, in_axes=(0, None))(poses, 1.0)
-        print(actions.shape)
-        new_poses = jax.vmap(self.env.batch_step, in_axes=(0, 0))(poses, actions)
-        return new_poses, new_poses
+            # Set the robot's velocity
+            dxu = np.array([[v], [w]])
 
-def main():
-    env = Robotarium(number_of_robots=1)
-    num_envs = 1
-    wrapped_env = WrappedRobotarium(env, num_envs)
-    initial_poses = jnp.zeros((num_envs, 3, 1))
-    final_poses, batch = jax.lax.scan(wrapped_env.batched_step, initial_poses, None, 1_000_000)
-    return batch
+            # Set the velocities of the robots
+            r.set_velocities(np.array([0]), dxu)
 
-if __name__ == "__main__":
-    run = jax.jit(main)
-    batch = jax.block_until_ready(run())
-    print(batch.shape)
+            # Iterate the simulation
+            r.step()
+            r.get_poses()
+            # poses.vstack((poses, r.get_poses()))
 
-    # Select one environment to plot
-    env_index = 0
-
-    # Extract x and y positions over timesteps
-    x_positions = batch[:, env_index, 0, 0]
-    y_positions = batch[:, env_index, 1, 0]
-
-    # Plot x and y positions
-    plt.figure(figsize=(5, 5))
-    plt.plot(x_positions, y_positions, label=f'Env {env_index}')
-    plt.xlabel('X Position')
-    plt.ylabel('Y Position')
-    plt.title('Trajectory of one environment over time')
-    plt.legend()
-    plt.grid(True)
-    plt.show()
+# Example usage
+num_env = 1
+num_t = 1_000_000
+drive_in_circle(num_env, num_t)
